requirements.txt
PyQt5==5.15.9
Pillow==9.5.0
requests==2.31.0
docx2txt==0.8
python-dotenv==1.0.0
torch==2.0.1
opencv-python==4.7.0.72
pytesseract==0.3.10
transformers==4.29.2
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
tokenizers==0.13.3
